Fix Mac M2 compatibility and GM MIDI specification
Local testing fixes:
- Fix Demucs detection for virtualenv installations
- Fix tempo validation (int to float for fractional BPM values)
- Correct GM MIDI bass range to 32-39 (was incorrectly 33-40)
- Correct GM MIDI drums range to 113-120 (was incorrectly 112-119)
- Update tests to match correct GM specifications

All tests passing (19/19)

Tested on Mac M2 8GB RAM with successful transcription

Assisted by Claude Code

+++ app/api/models.py
@@ -125,7 +125,7 @@ class TranscriptionResponse(BaseModel):
     job_id: str
     message: str
     duration: Optional[float] = Field(default=None, description="Audio duration in seconds")
-    tempo: Optional[int] = Field(default=None, description="Detected tempo in BPM")
+    tempo: Optional[float] = Field(default=None, description="Detected tempo in BPM")
     total_beats: Optional[int] = Field(default=None, description="Number of beats detected")
     stems_processed: Optional[int] = Field(default=None, description="Number of stems processed")
     total_segments: Optional[int] = Field(default=None, description="Total audio segments analyzed")
